Route stick-ball detection through the detector
